relocate system error from bin folder to error folder
